Update movie_remux.py
fix

+++ movie_remux.py
@@ -440,8 +440,9 @@ for file in os.listdir(movie_folder):
                                 subprocess.Popen(f'ffmpeg -i "{file1_path}"  -c:a pcm_s{bits}le -f w64 "{wav_file}"').wait()
                                 flac_file = os.path.splitext(file1_path)[0] + '.flac'
                                 subprocess.Popen(f'flac -8 -j {flac_threads} "{wav_file}" -o "{flac_file}"').wait()
-                                if os.path.getsize(flac_file) > os.path.getsize(file1_path):
-                                    os.remove(flac_file)
+                                if os.path.exists(flac_file):
+                                    if os.path.getsize(flac_file) > os.path.getsize(file1_path):
+                                        os.remove(flac_file)
                                 os.remove(file1_path)
                                 os.remove(wav_file)
                     flac_files = []
@@ -529,9 +530,10 @@ for file in os.listdir(movie_folder):
                                     subprocess.Popen(f'ffmpeg -i "{file1_path}"  -c:a pcm_s{bits}le -f w64 "{wav_file}"').wait()
                                     flac_file = os.path.splitext(file1_path)[0] + '.flac'
                                     subprocess.Popen(f'flac -8 -j {flac_threads} "{wav_file}" -o "{flac_file}"').wait()
-                                    if os.path.getsize(flac_file) > os.path.getsize(file1_path):
-                                        # 如果转换出来的flac体积比dts体积大则舍弃，这种情况很常见
-                                        os.remove(flac_file)
+                                    if os.path.exists(flac_file):
+                                        if os.path.getsize(flac_file) > os.path.getsize(file1_path):
+                                            # 如果转换出来的flac体积比dts体积大则舍弃，这种情况很常见
+                                            os.remove(flac_file)
                                     os.remove(file1_path)
                                     os.remove(wav_file)
                         flac_files = []
